Add get posts of followed members and get comments of followed members API calls
